wip: export aws/Application.on_*(..) functions to make them usable by DO
